fix: simplify activity feed real-time updates to use existing page socket
Previously, the activity sidebar had a separate broadcast system that:
- Required instrumentation.ts to set up hooks (wasn't executing in Docker)
- Used dedicated activity rooms (activity:drive:*, activity:page:*)
- Had its own broadcastActivityEvent function with debouncing

This simplifies to reuse the existing page tree socket:
- SidebarActivityTab now listens for page events (page:created, etc.)
- These events are already broadcast when pages change
- No additional broadcast infrastructure needed

Changes:
- Update SidebarActivityTab to use useSocket and listen for page events
- Remove broadcastActivityEvent calls from activity-logger.ts
- Delete useActivitySocket.ts hook (no longer needed)
- Delete instrumentation.ts (was setting up activity broadcast hook)
- Remove activity room handlers from realtime service
- Remove unused broadcast code from socket-utils.ts
- Update test mocks
